put withUpdateUser effect into one file
