--- App/Sagas/UserSagas/sendEmailVerification.js
@@ -1,8 +1,0 @@
-// @flow
-import { fork } from 'redux-saga/effects';
-import {sendEmailVerificationWithFirebase} from '../../Services/';
-import {withUpdateUser} from './index';
-
-export function* sendEmailVerification<T>(): Generator<T, any, any> {
-  yield fork(withUpdateUser, sendEmailVerificationWithFirebase);
-}
